task: Laravel + Vue

+++ src/database/migrations/2022_10_12_171648_create_products_table.php
@@ -27,7 +27,7 @@ return new class extends Migration
 
             $table->string('identifier')->unique();
 
-            $table->decimal('price', 10, 2)->nullable();
+            $table->integer('price')->nullable();
 
             $table->json('details')->nullable();
